Refactoring of skeleton hierarchy for local/global and ui/backend settings so that: -> implementations are contained neatly in nested classes -> the UISettings classes handle their own loading of both default settings and saved settings in their constructors -> the backend owns and maintains its own settings structure (backend settings will not even be cached in the UI layer).
Also, provide a stub Setting class, defined in the backend, from which both a BackendSetting and UISetting class is derived, and have the implementation classes simply maintain, internally, a vector of these Settings.  The Setting class must be copyable, for now.  Plan to use Boost.PropertyTree as temporary variables only to load XML from a file - NOT to store the data permanently in any Settings class.