The following tasks are implemented here (copied from https://basecamp.com/2097470/projects/1808931/messages/54026021#comment_448300879):
1) When an output file already exists, I think we want to ALSO ask the user if they wish to ``replace'' the file, not just append it.  Perhaps can can be rewritten to ask if the user if they wish to replace OR append to the existing dataset.  Then give the user the ability to click on a ``Replace'' button, an ``Append'' button, or a ``Cancel'' button.

Once they click on either the ``Replace'' or ``Append'' button, should we have a window that opens asking the user if they are sure?

2) In the DMU refresh screen, please change the the following text:

``Enter `code' column heading (data uniquely identifies each DMU member)''

to now read:

``Enter `id' column heading (uniquely identifies each DMU member)''

3) Is there a way to clear the ``Status'' window (not obvious how this is done)?

4) When a Variable group uses a UOA with multiple variable groups, is it possible for the columns # selection boxes to be listed in the order of the UOA?

For example, suppose the UOA is Country-ATOP Alliance.  In this case, if the user chooses ``ccode'' and ``atopid'', it would nice for the ``# COW Country cols:'' box to be listed BEFORE the ``#ATOP Alliance cols:'' box.  Presently, NewGene does the opposite.  I suspect this is because the boxes are listed alphabetically.